requirements.txt
pandas
numpy
streamlit
google-genai

flask
requests
beautifulsoup4
newspaper3k
lxml
lxml_html_clean
pymongo
plotly
